ACS-3015 move ErrorProneAction code to test folder
